services/agno-os/requirements.txt
agno[os]
fastapi
uvicorn
pydantic
python-dotenv
google-genai
